# perf: Reduce compile time by trimming template expansion in IBA