… src/include/OpenImageIO/imagebufalgo_util.h